headers
adding headers

+++ printf/ft_itoa_base.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_itoa_base.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ahugi <ahugi@student.42berlin.de>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/21 13:57:07 by ahugi             #+#    #+#             */
+/*   Updated: 2024/11/21 13:58:15 by ahugi            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "ft_printf.h"
 
 static int	ft_num_size(int n)
@@ -30,14 +42,14 @@ char	*ft_itoa_base(unsigned int n, int base)
 	number = (char *)ft_calloc((num_size + 1), sizeof(char));
 	if (!number)
 		return (NULL);
-	if (n == 0)
+	if ((int)n == 0)
 		number[0] = '0';
-	if (n < 0)
+	if ((int)n < 0)
 	{
 		number[0] = '-';
 		sign = -sign;
 	}
-	while (n != 0)
+	while ((int)n != 0)
 	{
 		number[num_size - 1] = base_symbols[(sign * n) % base];
 		n = n / base;
